♻️ [Refactoring]: flexbox.test.tsを機能別に分割
- colorsテストと同様の構造に変更
- __tests__ディレクトリを作成
- 機能別に9つのテストファイルに分割:
  - container: Flexコンテナ判定
  - direction: Flex方向
  - wrap: Flexラップ
  - justify: JustifyContent
  - align: AlignItems
  - gap: ギャップパース
  - spacing: スペーシングパース
  - padding: パディングパース
  - margin: マージンパース
- describeを削除しフラットな構造に変更
- テスト名を日本語に変更
- test.eachを使用してパラメータ化テストを導入
